Mireya/bot/requirements.txt
aiogram>=3.4.1
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
supabase>=2.4.0
python-dotenv>=1.0.1
aiohttp>=3.9.0
pydantic>=2.7.0
pandas>=2.1.0
matplotlib>=3.8.0
scikit-learn>=1.3.0
